psDraw.py: renamed and fixed to merge to master
